feat: 메인 페이지 index파일에서 MainPage로 변경
